Update to latest Scroom

+++ plugins/sep/seppresentation.hh
@@ -88,5 +88,5 @@ public:
   // PipetteViewInterface
   ////////////////////////////////////////////////////////////////////////
   PipetteLayerOperations::PipetteColor
-  getPixelAverages(Scroom::Utils::Rectangle<int> area) override;
+  getPixelAverages(Scroom::Utils::Rectangle<double> area) override;
 };
